# api/docs/swagger.yaml
definitions:
...
  models.Brand:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  models.CategorySwagger:
    properties:
      id:
        example: a2d70daf-b4ac-4198-a6a0-999447483c18
        type: string
      image:
        example: https://firebasestorage.googleapis.com/v0/b/ulab-market.appspot.com/o/test%2Fcategory%2Fa2d70daf-b4ac-4198-a6a0-999447483c18?alt=media&token=test%2Fcategory%2Fa2d70daf-b4ac-4198-a6a0-999447483c18
        type: string
      name:
        example: electronic
        type: string
      parent_id:
        example: ""
        type: string
      subcategories:
        items:
          $ref: '#/definitions/models.SubCategorySwagger'
        type: array
      translations:
        items:
          $ref: '#/definitions/models.CategoryTranslation'
        type: array
    type: object
  models.CategoryTranslation:
    properties:
      category_id:
        type: string
      language_code:
        type: string
      name:
        type: string
    type: object
...
  models.SubCategorySwagger:
    properties:
      id:
        type: string
      image:
        type: string
      name:
        type: string
      parent_id:
        type: string
      translations:
        items:
          $ref: '#/definitions/models.CategoryTranslation'
        type: array
    type: object
...
  models_v1.ChangeBrand:
    properties:
      id:
        type: string
      name:
        type: string
    required:
    - id
    - name
    type: object
  models_v1.ChangeCategory:
    properties:
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
    required:
    - id
    - name
    type: object
...
  models_v1.CreateBrand:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  models_v1.CreateCategory:
    properties:
      name:
        type: string
      parent_id:
        type: string
    required:
    - name
    type: object
...
  models_v1.Response:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  models_v1.ResponseID:
    properties:
      id:
        type: string
    type: object
...
host: localhost:8123
info:
  contact: {}
  title: E-commerce project v2
paths:
...
  /api/brand:
    get:
      consumes:
      - application/json
      description: get brand
      operationId: getAllBrand
      produces:
      - application/json
      responses:
        "200":
          description: brand returned
          schema:
            items:
              $ref: '#/definitions/models.Brand'
            type: array
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models_v1.Response'
      summary: get brand all
      tags:
      - brand
    post:
      consumes:
      - application/json
      description: Create brand
      operationId: createBrand
      parameters:
      - description: Create brand request
        in: body
        name: create_brand
        required: true
        schema:
          $ref: '#/definitions/models_v1.CreateBrand'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/models.Brand'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models_v1.Response'
        "409":
          description: Already exists
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/models_v1.Response'
      security:
      - ApiKeyAuth: []
      summary: Create brand
      tags:
      - brand
    put:
      consumes:
      - application/json
      description: change brand name
      operationId: changeBrand
      parameters:
      - description: change brand
        in: body
        name: changeBrand
        required: true
        schema:
          $ref: '#/definitions/models_v1.ChangeBrand'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/models_v1.Response'
        "400":
          description: Bad request / bad uuid / no update provided
          schema:
            $ref: '#/definitions/models_v1.Response'
        "404":
          description: Brand not found
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models_v1.Response'
      security:
      - ApiKeyAuth: []
      summary: change brand
      tags:
      - brand
  /api/brand/{id}:
    delete:
      consumes:
      - application/json
      description: delete brand
      operationId: deleteBrand
      parameters:
      - description: brand id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deleted successfully
          schema:
            $ref: '#/definitions/models_v1.Response'
        "400":
          description: bad uuid
          schema:
            $ref: '#/definitions/models_v1.Response'
        "404":
          description: brand not found
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models_v1.Response'
      security:
      - ApiKeyAuth: []
      summary: delete brand
      tags:
      - brand
    get:
      consumes:
      - application/json
      description: get brand by id
      operationId: getBrandById
      parameters:
      - description: brand id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: brand returned
          schema:
            $ref: '#/definitions/models.Brand'
        "400":
          description: Bad UUID
          schema:
            $ref: '#/definitions/models_v1.Response'
        "404":
          description: Brand not found
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models_v1.Response'
      summary: get brand by id
      tags:
      - brand
  /api/category:
    get:
      consumes:
      - application/json
      description: get category, returns translations, and subcategories for all category
      operationId: getAllCategory
      produces:
      - application/json
      responses:
        "200":
          description: category returned
          schema:
            items:
              $ref: '#/definitions/models.CategorySwagger'
            type: array
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models_v1.Response'
      summary: get category all
      tags:
      - category
    post:
      consumes:
      - application/json
      description: Create category
      operationId: createCategory
      parameters:
      - description: Create category request
        in: body
        name: create_category
        required: true
        schema:
          $ref: '#/definitions/models_v1.CreateCategory'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/models_v1.ResponseID'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models_v1.Response'
        "409":
          description: Already exists
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/models_v1.Response'
      security:
      - ApiKeyAuth: []
      summary: Create category
      tags:
      - category
    put:
      consumes:
      - application/json
      description: change category name and parent
      operationId: changeCategory
      parameters:
      - description: change category
        in: body
        name: changeCategory
        required: true
        schema:
          $ref: '#/definitions/models_v1.ChangeCategory'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/models_v1.Response'
        "400":
          description: Bad request / bad uuid
          schema:
            $ref: '#/definitions/models_v1.Response'
        "404":
          description: Category not found/ parent category not found
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models_v1.Response'
      security:
      - ApiKeyAuth: []
      summary: change category
      tags:
      - category
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"